+++ .gitignore
@@ -10,6 +10,8 @@
 /.idea/assetWizardSettings.xml
 /.idea/runConfigurations.xml
 /.idea/deploymentTargetSelector.xml
+/.idea/appInsightsSettings.xml
+/.idea/*.xml
 .DS_Store
 build/
 /captures
