+++ app/src/main/java/io/nekohasekai/sagernet/vpn/helpers/GenericHelper.kt
@@ -1,0 +1,462 @@
+package io.nekohasekai.sagernet.vpn.helpers
+
+class GenericHelper {
+    companion object {
+        fun countryMapper(countryCode: String): String {
+            val countries = mapOf(
+                "af" to "Afghanistan",
+                "al" to "Albania",
+                "dz" to "Algeria",
+                "as" to "American Samoa",
+                "ad" to "Andorra",
+                "ao" to "Angola",
+                "ai" to "Anguilla",
+                "ag" to "Antigua and Barbuda",
+                "ar" to "Argentina",
+                "am" to "Armenia",
+                "aw" to "Aruba",
+                "au" to "Australia",
+                "at" to "Austria",
+                "az" to "Azerbaijan",
+                "bs" to "Bahamas",
+                "bh" to "Bahrain",
+                "bd" to "Bangladesh",
+                "bb" to "Barbados",
+                "by" to "Belarus",
+                "be" to "Belgium",
+                "bz" to "Belize",
+                "bj" to "Benin",
+                "bm" to "Bermuda",
+                "bt" to "Bhutan",
+                "bo" to "Bolivia",
+                "ba" to "Bosnia and Herzegovina",
+                "bw" to "Botswana",
+                "br" to "Brazil",
+                "bn" to "Brunei",
+                "bg" to "Bulgaria",
+                "bf" to "Burkina Faso",
+                "bi" to "Burundi",
+                "cv" to "Cabo Verde",
+                "kh" to "Cambodia",
+                "cm" to "Cameroon",
+                "ca" to "Canada",
+                "ky" to "Cayman Islands",
+                "cf" to "Central African Republic",
+                "td" to "Chad",
+                "cl" to "Chile",
+                "cn" to "China",
+                "co" to "Colombia",
+                "km" to "Comoros",
+                "cd" to "Congo, Democratic Republic of the",
+                "cg" to "Congo, Republic of the",
+                "cr" to "Costa Rica",
+                "ci" to "Côte d'Ivoire",
+                "hr" to "Croatia",
+                "cu" to "Cuba",
+                "cy" to "Cyprus",
+                "cz" to "Czech Republic",
+                "dk" to "Denmark",
+                "dj" to "Djibouti",
+                "dm" to "Dominica",
+                "do" to "Dominican Republic",
+                "ec" to "Ecuador",
+                "eg" to "Egypt",
+                "sv" to "El Salvador",
+                "gq" to "Equatorial Guinea",
+                "er" to "Eritrea",
+                "ee" to "Estonia",
+                "et" to "Ethiopia",
+                "fk" to "Falkland Islands",
+                "fo" to "Faroe Islands",
+                "fj" to "Fiji",
+                "fi" to "Finland",
+                "fr" to "France",
+                "gf" to "French Guiana",
+                "pf" to "French Polynesia",
+                "ga" to "Gabon",
+                "gm" to "Gambia",
+                "ge" to "Georgia",
+                "de" to "Germany",
+                "gh" to "Ghana",
+                "gi" to "Gibraltar",
+                "gr" to "Greece",
+                "gd" to "Grenada",
+                "gp" to "Guadeloupe",
+                "gu" to "Guam",
+                "gt" to "Guatemala",
+                "gn" to "Guinea",
+                "gw" to "Guinea-Bissau",
+                "gy" to "Guyana",
+                "ht" to "Haiti",
+                "hn" to "Honduras",
+                "hu" to "Hungary",
+                "is" to "Iceland",
+                "in" to "India",
+                "id" to "Indonesia",
+                "ir" to "Iran",
+                "iq" to "Iraq",
+                "ie" to "Ireland",
+                "il" to "Israel",
+                "it" to "Italy",
+                "jm" to "Jamaica",
+                "jp" to "Japan",
+                "jo" to "Jordan",
+                "kz" to "Kazakhstan",
+                "ke" to "Kenya",
+                "ki" to "Kiribati",
+                "kp" to "North Korea",
+                "kr" to "South Korea",
+                "kw" to "Kuwait",
+                "kg" to "Kyrgyzstan",
+                "la" to "Laos",
+                "lv" to "Latvia",
+                "lb" to "Lebanon",
+                "ls" to "Lesotho",
+                "lr" to "Liberia",
+                "ly" to "Libya",
+                "li" to "Liechtenstein",
+                "lt" to "Lithuania",
+                "lu" to "Luxembourg",
+                "mk" to "North Macedonia",
+                "mg" to "Madagascar",
+                "mw" to "Malawi",
+                "my" to "Malaysia",
+                "mv" to "Maldives",
+                "ml" to "Mali",
+                "mt" to "Malta",
+                "mh" to "Marshall Islands",
+                "mr" to "Mauritania",
+                "mu" to "Mauritius",
+                "mx" to "Mexico",
+                "fm" to "Micronesia",
+                "md" to "Moldova",
+                "mc" to "Monaco",
+                "mn" to "Mongolia",
+                "me" to "Montenegro",
+                "ma" to "Morocco",
+                "mz" to "Mozambique",
+                "mm" to "Myanmar",
+                "na" to "Namibia",
+                "nr" to "Nauru",
+                "np" to "Nepal",
+                "nl" to "Netherlands",
+                "nc" to "New Caledonia",
+                "nz" to "New Zealand",
+                "ni" to "Nicaragua",
+                "ne" to "Niger",
+                "ng" to "Nigeria",
+                "nu" to "Niue",
+                "nf" to "Norfolk Island",
+                "mp" to "Northern Mariana Islands",
+                "no" to "Norway",
+                "om" to "Oman",
+                "pk" to "Pakistan",
+                "pw" to "Palau",
+                "pa" to "Panama",
+                "pg" to "Papua New Guinea",
+                "py" to "Paraguay",
+                "pe" to "Peru",
+                "ph" to "Philippines",
+                "pn" to "Pitcairn Islands",
+                "pl" to "Poland",
+                "pt" to "Portugal",
+                "pr" to "Puerto Rico",
+                "qa" to "Qatar",
+                "re" to "Réunion",
+                "ro" to "Romania",
+                "ru" to "Russia",
+                "rw" to "Rwanda",
+                "bl" to "Saint Barthélemy",
+                "sh" to "Saint Helena",
+                "kn" to "Saint Kitts and Nevis",
+                "lc" to "Saint Lucia",
+                "mf" to "Saint Martin",
+                "pm" to "Saint Pierre and Miquelon",
+                "vc" to "Saint Vincent and the Grenadines",
+                "ws" to "Samoa",
+                "sm" to "San Marino",
+                "sa" to "Saudi Arabia",
+                "sn" to "Senegal",
+                "rs" to "Serbia",
+                "sc" to "Seychelles",
+                "sl" to "Sierra Leone",
+                "sg" to "Singapore",
+                "sx" to "Sint Maarten",
+                "sk" to "Slovakia",
+                "si" to "Slovenia",
+                "sb" to "Solomon Islands",
+                "so" to "Somalia",
+                "za" to "South Africa",
+                "ss" to "South Sudan",
+                "es" to "Spain",
+                "lk" to "Sri Lanka",
+                "sd" to "Sudan",
+                "sr" to "Suriname",
+                "sz" to "Swaziland",
+                "se" to "Sweden",
+                "ch" to "Switzerland",
+                "sy" to "Syria",
+                "tw" to "Taiwan",
+                "tj" to "Tajikistan",
+                "tz" to "Tanzania",
+                "th" to "Thailand",
+                "tl" to "Timor-Leste",
+                "tg" to "Togo",
+                "tk" to "Tokelau",
+                "to" to "Tonga",
+                "tt" to "Trinidad and Tobago",
+                "tn" to "Tunisia",
+                "tr" to "Turkey",
+                "tm" to "Turkmenistan",
+                "tv" to "Tuvalu",
+                "ug" to "Uganda",
+                "ua" to "Ukraine",
+                "ae" to "United Arab Emirates",
+                "gb" to "United Kingdom",
+                "us" to "United States",
+                "uy" to "Uruguay",
+                "uz" to "Uzbekistan",
+                "vu" to "Vanuatu",
+                "va" to "Vatican City",
+                "ve" to "Venezuela",
+                "vn" to "Vietnam",
+                "wf" to "Wallis and Futuna",
+                "eh" to "Western Sahara",
+                "ye" to "Yemen",
+                "zm" to "Zambia",
+                "zw" to "Zimbabwe",
+                "other" to "Others"
+            )
+
+            return countries[countryCode].toString()
+        }
+
+        fun countryCodeMapper(countryName: String): String {
+            val countries = mapOf(
+                "Afghanistan" to "af",
+                "Albania" to "al",
+                "Algeria" to "dz",
+                "American Samoa" to "as",
+                "Andorra" to "ad",
+                "Angola" to "ao",
+                "Anguilla" to "ai",
+                "Antigua and Barbuda" to "ag",
+                "Argentina" to "ar",
+                "Armenia" to "am",
+                "Aruba" to "aw",
+                "Australia" to "au",
+                "Austria" to "at",
+                "Azerbaijan" to "az",
+                "Bahamas" to "bs",
+                "Bahrain" to "bh",
+                "Bangladesh" to "bd",
+                "Barbados" to "bb",
+                "Belarus" to "by",
+                "Belgium" to "be",
+                "Belize" to "bz",
+                "Benin" to "bj",
+                "Bermuda" to "bm",
+                "Bhutan" to "bt",
+                "Bolivia" to "bo",
+                "Bosnia and Herzegovina" to "ba",
+                "Botswana" to "bw",
+                "Brazil" to "br",
+                "Brunei" to "bn",
+                "Bulgaria" to "bg",
+                "Burkina Faso" to "bf",
+                "Burundi" to "bi",
+                "Cabo Verde" to "cv",
+                "Cambodia" to "kh",
+                "Cameroon" to "cm",
+                "Canada" to "ca",
+                "Cayman Islands" to "ky",
+                "Central African Republic" to "cf",
+                "Chad" to "td",
+                "Chile" to "cl",
+                "China" to "cn",
+                "Colombia" to "co",
+                "Comoros" to "km",
+                "Congo, Democratic Republic of the" to "cd",
+                "Congo, Republic of the" to "cg",
+                "Costa Rica" to "cr",
+                "Côte d'Ivoire" to "ci",
+                "Croatia" to "hr",
+                "Cuba" to "cu",
+                "Cyprus" to "cy",
+                "Czech Republic" to "cz",
+                "Denmark" to "dk",
+                "Djibouti" to "dj",
+                "Dominica" to "dm",
+                "Dominican Republic" to "do",
+                "Ecuador" to "ec",
+                "Egypt" to "eg",
+                "El Salvador" to "sv",
+                "Equatorial Guinea" to "gq",
+                "Eritrea" to "er",
+                "Estonia" to "ee",
+                "Ethiopia" to "et",
+                "Falkland Islands" to "fk",
+                "Faroe Islands" to "fo",
+                "Fiji" to "fj",
+                "Finland" to "fi",
+                "France" to "fr",
+                "French Guiana" to "gf",
+                "French Polynesia" to "pf",
+                "Gabon" to "ga",
+                "Gambia" to "gm",
+                "Georgia" to "ge",
+                "Germany" to "de",
+                "Ghana" to "gh",
+                "Gibraltar" to "gi",
+                "Greece" to "gr",
+                "Grenada" to "gd",
+                "Guadeloupe" to "gp",
+                "Guam" to "gu",
+                "Guatemala" to "gt",
+                "Guinea" to "gn",
+                "Guinea-Bissau" to "gw",
+                "Guyana" to "gy",
+                "Haiti" to "ht",
+                "Honduras" to "hn",
+                "Hungary" to "hu",
+                "Iceland" to "is",
+                "India" to "in",
+                "Indonesia" to "id",
+                "Iran" to "ir",
+                "Iraq" to "iq",
+                "Ireland" to "ie",
+                "Israel" to "il",
+                "Italy" to "it",
+                "Jamaica" to "jm",
+                "Japan" to "jp",
+                "Jordan" to "jo",
+                "Kazakhstan" to "kz",
+                "Kenya" to "ke",
+                "Kiribati" to "ki",
+                "North Korea" to "kp",
+                "South Korea" to "kr",
+                "Kuwait" to "kw",
+                "Kyrgyzstan" to "kg",
+                "Laos" to "la",
+                "Latvia" to "lv",
+                "Lebanon" to "lb",
+                "Lesotho" to "ls",
+                "Liberia" to "lr",
+                "Libya" to "ly",
+                "Liechtenstein" to "li",
+                "Lithuania" to "lt",
+                "Luxembourg" to "lu",
+                "North Macedonia" to "mk",
+                "Madagascar" to "mg",
+                "Malawi" to "mw",
+                "Malaysia" to "my",
+                "Maldives" to "mv",
+                "Mali" to "ml",
+                "Malta" to "mt",
+                "Marshall Islands" to "mh",
+                "Mauritania" to "mr",
+                "Mauritius" to "mu",
+                "Mexico" to "mx",
+                "Micronesia" to "fm",
+                "Moldova" to "md",
+                "Monaco" to "mc",
+                "Mongolia" to "mn",
+                "Montenegro" to "me",
+                "Morocco" to "ma",
+                "Mozambique" to "mz",
+                "Myanmar" to "mm",
+                "Namibia" to "na",
+                "Nauru" to "nr",
+                "Nepal" to "np",
+                "Netherlands" to "nl",
+                "New Caledonia" to "nc",
+                "New Zealand" to "nz",
+                "Nicaragua" to "ni",
+                "Niger" to "ne",
+                "Nigeria" to "ng",
+                "Niue" to "nu",
+                "Norfolk Island" to "nf",
+                "Northern Mariana Islands" to "mp",
+                "Norway" to "no",
+                "Oman" to "om",
+                "Pakistan" to "pk",
+                "Palau" to "pw",
+                "Panama" to "pa",
+                "Papua New Guinea" to "pg",
+                "Paraguay" to "py",
+                "Peru" to "pe",
+                "Philippines" to "ph",
+                "Pitcairn Islands" to "pn",
+                "Poland" to "pl",
+                "Portugal" to "pt",
+                "Puerto Rico" to "pr",
+                "Qatar" to "qa",
+                "Réunion" to "re",
+                "Romania" to "ro",
+                "Russia" to "ru",
+                "Rwanda" to "rw",
+                "Saint Barthélemy" to "bl",
+                "Saint Helena" to "sh",
+                "Saint Kitts and Nevis" to "kn",
+                "Saint Lucia" to "lc",
+                "Saint Martin" to "mf",
+                "Saint Pierre and Miquelon" to "pm",
+                "Saint Vincent and the Grenadines" to "vc",
+                "Samoa" to "ws",
+                "San Marino" to "sm",
+                "Saudi Arabia" to "sa",
+                "Senegal" to "sn",
+                "Serbia" to "rs",
+                "Seychelles" to "sc",
+                "Sierra Leone" to "sl",
+                "Singapore" to "sg",
+                "Sint Maarten" to "sx",
+                "Slovakia" to "sk",
+                "Slovenia" to "si",
+                "Solomon Islands" to "sb",
+                "Somalia" to "so",
+                "South Africa" to "za",
+                "South Sudan" to "ss",
+                "Spain" to "es",
+                "Sri Lanka" to "lk",
+                "Sudan" to "sd",
+                "Suriname" to "sr",
+                "Swaziland" to "sz",
+                "Sweden" to "se",
+                "Switzerland" to "ch",
+                "Syria" to "sy",
+                "Taiwan" to "tw",
+                "Tajikistan" to "tj",
+                "Tanzania" to "tz",
+                "Thailand" to "th",
+                "Timor-Leste" to "tl",
+                "Togo" to "tg",
+                "Tokelau" to "tk",
+                "Tonga" to "to",
+                "Trinidad and Tobago" to "tt",
+                "Tunisia" to "tn",
+                "Turkey" to "tr",
+                "Turkmenistan" to "tm",
+                "Tuvalu" to "tv",
+                "Uganda" to "ug",
+                "Ukraine" to "ua",
+                "United Arab Emirates" to "ae",
+                "United Kingdom" to "gb",
+                "United States" to "us",
+                "Uruguay" to "uy",
+                "Uzbekistan" to "uz",
+                "Vanuatu" to "vu",
+                "Vatican City" to "va",
+                "Venezuela" to "ve",
+                "Vietnam" to "vn",
+                "Wallis and Futuna" to "wf",
+                "Western Sahara" to "eh",
+                "Yemen" to "ye",
+                "Zambia" to "zm",
+                "Zimbabwe" to "zw",
+                "Others" to "other"
+            )
+            return countries[countryName].toString()
+        }
+    }
+}
