+++ app/src/main/res/layout/layout_progress_list.xml
@@ -15,6 +15,6 @@
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/list_view"
         android:layout_width="match_parent"
-        android:layout_height="match_parent" />
+        android:layout_height="80dp" />
 
 </LinearLayout>
